Stack
just need to find the breaking point of the increasing order and then removing digits if it's not greater than the stack top. Once max number of reduction is reached just push everything into the stack. Sometimes  then may be digits present in the string not removed in the first attempt we should then pop elements until k>0 and then return a substring of the partially filled up string with elements of the stack